Move logics to main_logics.rs to make main.rs slim
