+++ base/src/main.rs
@@ -1,9 +1,9 @@
+mod main_logics;
+
 use dotenv;
-use reqwest;
+use main_logics::post_current_song;
 use twitter_api::api_model::api::Api as TwitterAPI;
 use spotify_api::api_model::api::Api as SpotifyAPI;
-use spotify_api::api_model::responses::CurrentSong;
-use interface::PostAPI;
 use std::io::stdin;
 use std::io::{stdout, Write};
 
@@ -36,80 +36,6 @@ async fn get_spotify_api(cred: &AuthKey) -> SpotifyAPI {
         }
         Err(reason) => {
             panic!("{}", reason)
-        }
-    }
-}
-
-async fn post_current_song_on_local(
-    data: CurrentSong,
-    papi: &dyn PostAPI
-) -> Result<(), String> {
-    let artists = &data.track_artists.join(", ");
-    let text = if &artists.chars().count() > &0 {
-        format!{"#nowplaying {} - {}", &data.song_title, artists}
-    }
-    else {
-        format!{"#nowplaying {}", &data.song_title}
-    };
-
-    match papi.compose_without_picture(&text).await {
-        Ok(_) => Ok(()),
-        Err(_) => Err("Tweet failed.".to_string())
-    }
-}
-
-async fn post_current_song_on_spotify(
-    data: CurrentSong,
-    papi: &dyn PostAPI
-) -> Result<(), String> {
-    let img = reqwest::get(&data.album_art_url.to_owned().unwrap()).await
-        .unwrap()
-        .bytes().await
-        .unwrap();
-
-    match papi.upload_media(img).await {
-        Some(id) => {
-            let media_ids = vec![id];
-            let song_url = format!{"https://open.spotify.com/track/{}", &data.song_uri.to_owned().unwrap()};
-            let artists = &data.track_artists.join(", ");
-            let text = format!{
-                "#nowplaying {} - {}\n{}",
-                &data.song_title,
-                artists,
-                song_url
-            };
-
-            let _ = papi.compose_with_picture(&text, &media_ids).await;
-            Ok(())
-        }
-        None => {
-            Err("Tweet failed. Could not upload a image to twitter".to_string())
-        }
-    }
-}
-
-async fn post_current_song(
-    papi: &dyn PostAPI,
-    sapi: &mut SpotifyAPI<'_>
-) -> Result<(), String> {
-    match sapi.fetch_current_song().await {
-        Ok(raw) => {
-            let resp = sapi.parse_current_song_result(raw);
-
-            let post_result = if resp.album_art_url.is_none() {
-                post_current_song_on_local(resp, papi).await
-            }
-            else {
-                post_current_song_on_spotify(resp, papi).await
-            };
-
-            match post_result {
-                Ok(_) => Ok(()),
-                Err(reason) => Err(reason.to_string())
-            }
-        }
-        Err(reason) => {
-            Err(reason.to_string())
         }
     }
 }
